adjust project structure and rename files
