+++ src/app/auth/sign-in/page.tsx
@@ -2,9 +2,9 @@ import * as React from "react";
 import type { Metadata } from "next";
 
 import { config } from "@/config";
-import { GuestGuard } from "@/components/auth/guest-guard";
+import { GuestGuard } from "@/components/auth/GuestGuard";
 import { Layout } from "@/components/auth/layout";
-import { SignInForm } from "@/components/auth/sign-in-form";
+import { SignInForm } from "@/components/auth/SignInForm";
 
 export const metadata = {
   title: `Sign in | Auth | ${config.site.name}`,
