+++ src/app/auth/sign-up/page.tsx
@@ -2,9 +2,9 @@ import * as React from "react";
 import type { Metadata } from "next";
 
 import { config } from "@/config";
-import { GuestGuard } from "@/components/auth/guest-guard";
+import { GuestGuard } from "@/components/auth/GuestGuard";
 import { Layout } from "@/components/auth/layout";
-import { SignUpForm } from "@/components/auth/sign-up-form";
+import { SignUpForm } from "@/components/auth/SignUpForm";
 
 export const metadata = {
   title: `Sign up | Auth | ${config.site.name}`,
