+++ src/components/auth/SignInForm.tsx
@@ -3,51 +3,45 @@
 import * as React from "react";
 import RouterLink from "next/link";
 import { useRouter } from "next/navigation";
+import { Controller, useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
+import { z as zod } from "zod";
+
 import Alert from "@mui/material/Alert";
 import Button from "@mui/material/Button";
-import Checkbox from "@mui/material/Checkbox";
 import FormControl from "@mui/material/FormControl";
-import FormControlLabel from "@mui/material/FormControlLabel";
 import FormHelperText from "@mui/material/FormHelperText";
 import InputLabel from "@mui/material/InputLabel";
 import Link from "@mui/material/Link";
 import OutlinedInput from "@mui/material/OutlinedInput";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import { Controller, useForm } from "react-hook-form";
-import { z as zod } from "zod";
+
+import { Eye as EyeIcon } from "@phosphor-icons/react/dist/ssr/Eye";
+import { EyeSlash as EyeSlashIcon } from "@phosphor-icons/react/dist/ssr/EyeSlash";
 
 import { paths } from "@/paths";
 import { authClient } from "@/lib/auth/client";
-import { useUser } from "@/hooks/use-user";
+import { useUser } from "@/hooks/useUser";
 
 const schema = zod.object({
-  firstName: zod.string().min(1, { message: "First name is required" }),
-  lastName: zod.string().min(1, { message: "Last name is required" }),
   email: zod.string().min(1, { message: "Email is required" }).email(),
-  password: zod
-    .string()
-    .min(6, { message: "Password should be at least 6 characters" }),
-  terms: zod
-    .boolean()
-    .refine((value) => value, "You must accept the terms and conditions"),
+  password: zod.string().min(1, { message: "Password is required" }),
 });
 
 type Values = zod.infer<typeof schema>;
 
 const defaultValues = {
-  firstName: "",
-  lastName: "",
   email: "",
   password: "",
-  terms: false,
 } satisfies Values;
 
-export function SignUpForm(): React.JSX.Element {
+export function SignInForm(): React.JSX.Element {
   const router = useRouter();
 
   const { checkSession } = useUser();
+
+  const [showPassword, setShowPassword] = React.useState<boolean>();
 
   const [isPending, setIsPending] = React.useState<boolean>(false);
 
@@ -62,7 +56,7 @@ export function SignUpForm(): React.JSX.Element {
     async (values: Values): Promise<void> => {
       setIsPending(true);
 
-      const { error } = await authClient.signUp(values);
+      const { error } = await authClient.signIn(values);
 
       if (error) {
         setError("root", { type: "server", message: error });
@@ -81,49 +75,23 @@ export function SignUpForm(): React.JSX.Element {
   );
 
   return (
-    <Stack spacing={3}>
+    <Stack spacing={4}>
       <Stack spacing={1}>
-        <Typography variant="h4">Sign up</Typography>
+        <Typography variant="h4">Sign in</Typography>
         <Typography color="text.secondary" variant="body2">
-          Already have an account?{" "}
+          Don&apos;t have an account?{" "}
           <Link
             component={RouterLink}
-            href={paths.auth.signIn}
+            href={paths.auth.signUp}
             underline="hover"
             variant="subtitle2"
           >
-            Sign in
+            Sign up
           </Link>
         </Typography>
       </Stack>
       <form onSubmit={handleSubmit(onSubmit)}>
         <Stack spacing={2}>
-          <Controller
-            control={control}
-            name="firstName"
-            render={({ field }) => (
-              <FormControl error={Boolean(errors.firstName)}>
-                <InputLabel>First name</InputLabel>
-                <OutlinedInput {...field} label="First name" />
-                {errors.firstName ? (
-                  <FormHelperText>{errors.firstName.message}</FormHelperText>
-                ) : null}
-              </FormControl>
-            )}
-          />
-          <Controller
-            control={control}
-            name="lastName"
-            render={({ field }) => (
-              <FormControl error={Boolean(errors.firstName)}>
-                <InputLabel>Last name</InputLabel>
-                <OutlinedInput {...field} label="Last name" />
-                {errors.firstName ? (
-                  <FormHelperText>{errors.firstName.message}</FormHelperText>
-                ) : null}
-              </FormControl>
-            )}
-          />
           <Controller
             control={control}
             name="email"
@@ -143,37 +111,50 @@ export function SignUpForm(): React.JSX.Element {
             render={({ field }) => (
               <FormControl error={Boolean(errors.password)}>
                 <InputLabel>Password</InputLabel>
-                <OutlinedInput {...field} label="Password" type="password" />
+                <OutlinedInput
+                  {...field}
+                  endAdornment={
+                    showPassword ? (
+                      <EyeIcon
+                        cursor="pointer"
+                        fontSize="var(--icon-fontSize-md)"
+                        onClick={(): void => {
+                          setShowPassword(false);
+                        }}
+                      />
+                    ) : (
+                      <EyeSlashIcon
+                        cursor="pointer"
+                        fontSize="var(--icon-fontSize-md)"
+                        onClick={(): void => {
+                          setShowPassword(true);
+                        }}
+                      />
+                    )
+                  }
+                  label="Password"
+                  type={showPassword ? "text" : "password"}
+                />
                 {errors.password ? (
                   <FormHelperText>{errors.password.message}</FormHelperText>
                 ) : null}
               </FormControl>
             )}
           />
-          <Controller
-            control={control}
-            name="terms"
-            render={({ field }) => (
-              <div>
-                <FormControlLabel
-                  control={<Checkbox {...field} />}
-                  label={
-                    <React.Fragment>
-                      I have read the <Link>terms and conditions</Link>
-                    </React.Fragment>
-                  }
-                />
-                {errors.terms ? (
-                  <FormHelperText error>{errors.terms.message}</FormHelperText>
-                ) : null}
-              </div>
-            )}
-          />
+          <div>
+            <Link
+              component={RouterLink}
+              href={paths.auth.resetPassword}
+              variant="subtitle2"
+            >
+              Forgot password?
+            </Link>
+          </div>
           {errors.root ? (
             <Alert color="error">{errors.root.message}</Alert>
           ) : null}
           <Button disabled={isPending} type="submit" variant="contained">
-            Sign up
+            Sign in
           </Button>
         </Stack>
       </form>
