+++ src/components/dashboard/layout/MainNav.tsx
@@ -7,7 +7,7 @@ import IconButton from "@mui/material/IconButton";
 import Stack from "@mui/material/Stack";
 import { List as ListIcon } from "@phosphor-icons/react/dist/ssr/List";
 
-import { usePopover } from "@/hooks/use-popover";
+import { usePopover } from "@/hooks/usePopover";
 
 import { MobileNav } from "./MobileNav";
 import { UserPopover } from "./UserPopover";
