+++ src/components/dashboard/layout/SideNav.tsx
@@ -13,12 +13,12 @@ import { ArrowSquareUpRight as ArrowSquareUpRightIcon } from "@phosphor-icons/re
 import type { NavItemConfig } from "@/types/nav";
 import { paths } from "@/paths";
 import { isNavItemActive } from "@/lib/is-nav-item-active";
-import { Logo } from "@/components/core/logo";
+import { Logo } from "@/components/core/Logo";
 
 import { navItems } from "./config";
 import { navIcons } from "./NavIcons";
 import { authClient } from "@/lib/auth/client";
-import { useUser } from "@/hooks/use-user";
+import { useUser } from "@/hooks/useUser";
 import { logger } from "@/lib/default-logger";
 
 export function SideNav(): React.JSX.Element {
