+++ src/components/dashboard/overview/ItemStock.tsx
@@ -14,7 +14,7 @@ import { DeviceTablet as DeviceTabletIcon } from "@phosphor-icons/react/dist/ssr
 import { Phone as PhoneIcon } from "@phosphor-icons/react/dist/ssr/Phone";
 import type { ApexOptions } from "apexcharts";
 
-import { Chart } from "@/components/core/chart";
+import { Chart } from "@/components/core/Chart";
 
 const iconMapping = {
   Desktop: DesktopIcon,
