+++ src/hooks/useUser.ts
@@ -1,7 +1,7 @@
 import * as React from "react";
 
-import type { UserContextValue } from "@/contexts/user-context";
-import { UserContext } from "@/contexts/user-context";
+import type { UserContextValue } from "@/contexts/UserContext";
+import { UserContext } from "@/contexts/UserContext";
 
 export function useUser(): UserContextValue {
   const context = React.useContext(UserContext);
